Donut fix. Needed a cast.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@5966 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/util/xp/ut_AbiObject.cpp
@@ -105,7 +105,7 @@ bool UT_AbiObject::equal (UT_AbiObject * other) const
 UT_uint32 UT_AbiObject::hashcode () const
 {
   // 9987001 is a reasnonably large prime
-  long theAddr = reinterpret_cast<long>((this) * 0x9863b9);
+  long theAddr = reinterpret_cast<long>((long)(this) * 0x9863b9);
   return (UT_uint32) (theAddr);
 }
